Revises table creation script
The database table creation/population script previously did *both*
the creation of the tables and the addition of sample data. For
creating a production instance, we want these actions to be
independent (i.e. you may want to create the tables, but populate them
with *real* data).

This patch makes it possible to do either action independently or both
in sequence.